// frontend/generated/models/VoucherDto.ts
export interface VoucherDto {
    /**
     * 
     * @type {number}
     * @memberof VoucherDto
     */
    voucherId?: number;
    /**
     * 
     * @type {number}
     * @memberof VoucherDto
     */
    orderId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof VoucherDto
     */
    account?: AccountDto;
    /**
     * 
     * @type {string}
     * @memberof VoucherDto
     */
    code?: string;
    /**
     * 
     * @type {number}
     * @memberof VoucherDto
     */
    discountRate?: number;
    /**
     * 
     * @type {number}
     * @memberof VoucherDto
     */
    limitAmount?: number;
    /**
     * 
     * @type {string}
     * @memberof VoucherDto
     */
    state?: VoucherDtoStateEnum;
    /**
     * 
     * @type {Date}
     * @memberof VoucherDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof VoucherDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof VoucherDto
     */
    expiredAt?: Date;
}


/**
 * @export
 */
export const VoucherDtoStateEnum = {
    Used: 'USED',
    Available: 'AVAILABLE',
    Reserved: 'RESERVED',
    Expired: 'EXPIRED'
} as const;
export type VoucherDtoStateEnum = typeof VoucherDtoStateEnum[keyof typeof VoucherDtoStateEnum];